Update _index.md
we can give an option which is already available now for the learners to go and automatically build the spring/camunda project easily and quickly. This will give log of comfort to beginners to start with and focus more on learning camunda specifics rather than spending time in setting up the project.